Improve the FramerAPI resolve() method
When provided a type hint we no longer fallback to looking up any type.
This previously would result in a namespace being returned when
expecting a function. Instead we're more lenient about the different
types we support, for example, we will now return a MethodSignature type
when passing the Method type hint. This means in most cases the function
should return the object you were expecting.

The verify-api-references command has also had an overhaul, it will
now print out the names of the models that are missing rather than
just the filename and count.